requirements.txt
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap4==4.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
python-dotenv==0.10.3
requests==2.22.0
soupsieve==1.9.5
SQLAlchemy==1.3.12
urllib3==1.25.7
visitor==0.1.3
Werkzeug==0.16.0
